+++ mongeul/src/app/layout.tsx
@@ -1,11 +1,11 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import Header from "@/components/Header";
-import Navbar from "@/components/NavBar";
+import Header from "@/components/header/orgamisms/Header";
+import Navbar from "@/components/navbar/organisms/NavBar";
 import { ThemeProvider } from "@/context/ThemeContext";
 import { Provider } from "react-redux";
 import { store } from "@/store/store";
-import ClientLayout from "@/components/ClientLayout";
+import ClientLayout from "@/components/layout/ClientLayout";
 
 export const metadata: Metadata = {
   title: "몽글몽글",
