+++ mongeul/src/components/common/atoms/Textarea.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import Card from "../../Card";
+import Card from "./Card";
 import clsx from "clsx";
 
 interface TextareaProps {
