+++ mongeul/src/components/layout/ParellelLayout.tsx
@@ -13,10 +13,10 @@ export default function ParellelLayoutProps({
   // TODO 모바일 반응형 children / detail 하나만 보이기, 헤더 뒤로가기 추가
   return (
     <div className="flex gap-4 w-full justify-center">
-      <div className="w-full md:w-1/2">
+      <div className="w-full md:w-1/2 hidden md:block">
         <DefaultLayout>{children}</DefaultLayout>
       </div>
-      <div className="w-full md:w-1/2">
+      <div className="w-full md:w-1/2 block">
         <DefaultLayout>{detail}</DefaultLayout>
       </div>
     </div>
