+++ mongeul/src/components/write-diary/molecules/FeelingsSelect.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import { useState } from "react";
-import RoundIcon from "../atoms/RoundIcon";
-import MobileModal from "../atoms/MobileModal";
+import RoundIcon from "../../common/atoms/RoundIcon";
+import MobileModal from "../../common/atoms/MobileModal";
 import FeelingsIcon from "@/assets/icons/feelings.svg";
 import HappyIcon from "@/assets/icons/happy.svg";
-import WebModal from "../atoms/WebModal";
+import WebModal from "../../common/atoms/WebModal";
 import { RootState } from "@/store/store";
 import { useDispatch, useSelector } from "react-redux";
 import { setFeelings } from "@/store/diarySlice";
